split out the creature from the creatures.
