+++ aspirant_application/Game.Creature.h
@@ -1,4 +1,7 @@
 #pragma once
+#include "Common.XY.h"
+#include <optional>
+#include <string>
 namespace game
 {
 	enum class Creature
@@ -12,4 +15,20 @@ namespace game
 		WEREWOLF,
 		WYVERN
 	};
+
+}
+namespace game::creature
+{
+	struct Descriptor
+	{
+		std::string imageId;
+		int maximumHealth;
+		int attack;
+		int defend;
+		int foodBribe;
+		int moneyBribe;
+	};
+
+	Descriptor GetDescriptor(game::Creature);
+	void InitializeFromFile(const std::string&);
 }
