+++ aspirant_application/Game.Creatures.h
@@ -3,22 +3,6 @@
 #include "Common.XY.h"
 #include <optional>
 #include <string>
-namespace game::creature
-{
-	struct Descriptor
-	{
-		std::string imageId;
-		int maximumHealth;
-		int attack;
-		int defend;
-		int foodBribe;
-		int moneyBribe;
-	};
-
-	Descriptor GetDescriptor(game::Creature);
-	std::optional<Descriptor> GetDescriptor(const common::XY<size_t>&);
-	void InitializeFromFile(const std::string&);
-}
 namespace game::Creatures
 {
 	struct Instance
