+++ aspirant_application/aspirant_application.vcxproj
@@ -174,6 +174,7 @@
     <ClCompile Include="Game.Combat.cpp" />
     <ClCompile Include="Game.CombatDeck.cpp" />
     <ClCompile Include="Game.cpp" />
+    <ClCompile Include="Game.Creature.cpp" />
     <ClCompile Include="Game.Creatures.cpp" />
     <ClCompile Include="Game.Items.cpp" />
     <ClCompile Include="Game.World.cpp" />
