+++ aspirant_application/aspirant_application.vcxproj.filters
@@ -215,6 +215,9 @@
     <ClCompile Include="Graphics.CardSprites.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="Game.Creature.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Common.RNG.h">
